Primitive Input/Handling
Extremely basic input handling and error messaging for deposits/withdrawals/creating passwords, no data storing or passing between activities yet.

+++ BankingApp/app/src/main/java/com/example/bankingapp/DepositActivity.kt
@@ -2,6 +2,8 @@ package com.example.bankingapp
 
 import android.os.Bundle
 import android.widget.Button
+import android.widget.EditText
+import android.widget.Toast
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 
@@ -11,19 +13,21 @@ class DepositActivity : AppCompatActivity() {
         enableEdgeToEdge()
         setContentView(R.layout.activity_deposit)
 
-        //val activityLauncher = registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { _ -> }
-        // Listener for Submit Deposit Button
+        val depositAmount = findViewById<EditText>(R.id.DepositAmount)
         findViewById<Button>(R.id.SubmitDeposit).setOnClickListener{
             // assume that user has sufficient funds to make deposit
+            val depositValue = depositAmount.text.toString().toDoubleOrNull()
             // ~~~(evaluate and submit deposit)~~~
-
-            // Launcher above and Intent Below could be used for separate confirmation screen
-            // but i think it'll be easier to just return to ChooseAction and display a message with...
-            finish() // kill the activity and return to ChooseActionActivity
-
-            //val confirmIntent = Intent(this, ConfirmationActivity::class.java)
-            //activityLauncher.launch(confirmIntent) //
-
+            if (depositValue != null){
+                if (depositValue in 0.01..5000.0) {
+                    //~~~~~(add deposit amount to account balance)~~~~~
+                    //~~~~~(send confirmation message back to ChooseActionActivity)~~~~~
+                    finish() // kill the activity and return to ChooseActionActivity
+                }
+                else if (5000.0 < depositValue) Toast.makeText(this, "Single deposit limit is $5,000.00", Toast.LENGTH_SHORT).show()
+                else Toast.makeText(this, "Deposit minimum is $0.01", Toast.LENGTH_SHORT).show()
+            }
+            else Toast.makeText(this, "Deposit minimum is $0.01", Toast.LENGTH_SHORT).show()
         }
 
     } // End of onCreate()
